feat(budget): the budget tab reads as a receipt — card, hugged pencil, demoted suggestion
Wave-1 budget-tab redesign (TripAdvisor/Wanderlog/Tripsy references):

- Extract the add-expense composer to widgets/trip_detail/budget_composer.dart
  as a PURE MOVE (approved): controllers + draft write-back + measured add
  row + add-options line travel verbatim; the add still goes through the
  section's _run. No visual change to the composer.
- Receipt card: groups + totals now live on ONE raised surface (card theme),
  with the Group-by toggle moved into the card's head row next to what it
  acts on (its 'Group by' label retires; the card head names the contents).
- The target pencil hugs the headline (and the empty-state target row)
  instead of hanging at the page's far edge.
- Group headers take a tinted icon tile (24px) and a w600 label so they
  lead their rows.
- Totals: 'Total planned' drops to the memo register (muted w600); 'Total
  spent' stays the one emphasized total — TripAdvisor's receipt discipline.
- The daily food & drink suggestion moves into a recessed well
  (surfaceContainerHighest, no shadow) that collapses with the section, so
  the labelled estimate reads as the secondary surface, not a rival receipt.
- New l10n key budgetExpensesTitle (Expenses / Gastos); arbs edited by text
  insertion, codegen via flutter gen-l10n after pub get repaired its
  container-baked pub-cache path.

Row state circles, kebab menus (#490 register), and dialogs are unchanged:
the circles are pinned by budget_amounts_test.dart.
