+++ src/packages/flutter-app/lib/l10n/app_en.arb
@@ -463,6 +463,7 @@
   "budgetGroupBy": "Group by",
   "budgetGroupByCategory": "Category",
   "budgetGroupByCity": "City",
+  "budgetExpensesTitle": "Expenses",
   "budgetGroupRestOfTrip": "Rest of trip",
   "budgetCityLabel": "City",
   "budgetCityNone": "No city",
