+++ src/packages/flutter-app/lib/l10n/app_es.arb
@@ -307,6 +307,7 @@
   "budgetGroupBy": "Agrupar por",
   "budgetGroupByCategory": "Categoría",
   "budgetGroupByCity": "Ciudad",
+  "budgetExpensesTitle": "Gastos",
   "budgetGroupRestOfTrip": "Resto del viaje",
   "budgetCityLabel": "Ciudad",
   "budgetCityNone": "Sin ciudad",
