+++ src/packages/flutter-app/lib/l10n/app_localizations.dart
@@ -1940,6 +1940,12 @@ abstract class AppLocalizations {
   /// **'City'**
   String get budgetGroupByCity;
 
+  /// No description provided for @budgetExpensesTitle.
+  ///
+  /// In en, this message translates to:
+  /// **'Expenses'**
+  String get budgetExpensesTitle;
+
   /// No description provided for @budgetGroupRestOfTrip.
   ///
   /// In en, this message translates to:
