+++ src/packages/flutter-app/lib/l10n/app_localizations_en.dart
@@ -995,6 +995,9 @@ class AppLocalizationsEn extends AppLocalizations {
   String get budgetGroupByCity => 'City';
 
   @override
+  String get budgetExpensesTitle => 'Expenses';
+
+  @override
   String get budgetGroupRestOfTrip => 'Rest of trip';
 
   @override
